# backend/contracts/openapi.yaml
{
  "components": {
    "schemas": {
...
      "StrategyRequest": {
        "additionalProperties": false,
        "properties": {
          "custom_words_limit": {
            "default": 5,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Custom Words Limit",
            "type": "integer"
          },
          "due_words_limit": {
            "default": 5,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Due Words Limit",
            "type": "integer"
          },
          "error_words_limit": {
            "default": 5,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Error Words Limit",
            "type": "integer"
          },
          "fallback_unreviewed_days": {
            "default": 3,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Fallback Unreviewed Days",
            "type": "integer"
          },
          "new_words_limit": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "New Words Limit",
            "type": "integer"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "total_words": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Words"
          },
          "word_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 200,
            "title": "Word Ids",
            "type": "array"
          }
        },
        "title": "StrategyRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "opaque",
        "description": "Managed high-entropy bearer token. Required scopes are listed in x-required-scopes because this is not an OAuth token endpoint.",
        "scheme": "bearer",
        "type": "http",
        "x-available-scopes": [
          "practice:generate",
          "practice:read",
          "practice:write",
          "reviews:read",
          "reviews:write",
          "words:export",
          "words:read",
          "words:write"
        ]
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "TrustedProxyUser": {
        "description": "Only accepted from TRUSTED_PROXY_CIDRS after the proxy strips client-supplied headers.",
        "in": "header",
        "name": "X-Forwarded-User",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-user word memory API. Browser requests use a trusted reverse-proxy identity; external Skills use scoped bearer tokens.",
    "title": "Word Memory Assistant API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/daily-table/generate": {
      "post": {
        "operationId": "generate_api_v1_daily_table_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Generate",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:generate"
        ]
      }
    },
    "/api/v1/practice-review-rounds/{round_id}": {
      "get": {
        "operationId": "get_round_api_v1_practice_review_rounds__round_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Get Round",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:read"
        ]
      }
    },
    "/api/v1/practice-review-rounds/{round_id}/items/{item_id}/result": {
      "put": {
        "operationId": "put_result_api_v1_practice_review_rounds__round_id__items__item_id__result_put",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Put Result",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "reviews:write"
        ]
      }
    },
    "/api/v1/practice-review-rounds/{round_id}/results": {
      "put": {
        "operationId": "put_results_api_v1_practice_review_rounds__round_id__results_put",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchResults"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Put Results",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "reviews:write"
        ]
      }
    },
    "/api/v1/practice-sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_practice_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "created_by_actor_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created By Actor Type"
            }
          },
          {
            "in": "query",
            "name": "generated_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Generated From"
            }
          },
          {
            "in": "query",
            "name": "generated_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Generated To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:read"
        ]
      }
    },
    "/api/v1/practice-sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_route_api_v1_practice_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Expected Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Delete Session Route",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:write"
        ]
      },
      "get": {
        "operationId": "get_session_api_v1_practice_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:read"
        ]
      },
      "patch": {
        "operationId": "update_session_route_api_v1_practice_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Update Session Route",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:write"
        ]
      }
    },
    "/api/v1/practice-sessions/{session_id}/archive": {
      "post": {
        "operationId": "archive_api_v1_practice_sessions__session_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Archive",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:write"
        ]
      }
    },
...
  }
}